Split API Model into separate classes
