+++ Jellyfin.Plugin.Kitsu/Providers/KitsuIO/ApiClient/Models/Included.cs
@@ -1,0 +1,7 @@
+﻿namespace Jellyfin.Plugin.Kitsu.Providers.KitsuIO.ApiClient.Models
+{
+    public class Included
+    {
+        public IncludedAttributes Attributes { get; set; }
+    }
+}
