+++ Jellyfin.Plugin.Kitsu/Providers/KitsuIO/ApiClient/Models/IncludedAttributes.cs
@@ -1,0 +1,7 @@
+﻿namespace Jellyfin.Plugin.Kitsu.Providers.KitsuIO.ApiClient.Models
+{
+    public class IncludedAttributes
+    {
+        public string Name { get; set; }
+    }
+}
