+++ Jellyfin.Plugin.Kitsu/Providers/KitsuIO/ApiClient/Models/KitsuEpisode.cs
@@ -1,21 +1,8 @@
-using System;
-
 namespace Jellyfin.Plugin.Kitsu.Providers.KitsuIO.ApiClient.Models
 {
     public class KitsuEpisode
     {
         public long Id { get; set; }
         public KitsuEpisodeAttributes Attributes { get; set; }
-    }
-
-    public class KitsuEpisodeAttributes
-    {
-        public string Synopsis { get; set; }
-        public KitsuTitles Titles { get; set; }
-        public int? Number { get; set; }
-        public int? SeasonNumber { get; set; }
-        public DateTime? AirDate { get; set; }
-        public int? Length { get; set; }
-        public KitsuImage Thumbnail { get; set; }
     }
 }
