+++ Jellyfin.Plugin.Kitsu/Providers/KitsuIO/ApiClient/Models/ResponseMeta.cs
@@ -1,0 +1,7 @@
+﻿namespace Jellyfin.Plugin.Kitsu.Providers.KitsuIO.ApiClient.Models
+{
+    public class ResponseMeta
+    {
+        public long? Count { get; set; }
+    }
+}
